UML3599 restructured code | wip
