+++ service-api/app/src/App/src/Service/Lpa/IsValid/IsValidInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Service\Lpa\IsValid;
+
+interface IsValidInterface
+{
+    public function getStatus(): string;
+
+    public function getUid(): string;
+}
