--- service-api/app/src/App/src/Service/Lpa/IsValid/LpaValidationInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Service\Lpa\IsValid;
-
-use App\DataAccess\ApiGateway\SiriusLpas;
-
-interface LpaValidationInterface
-{
-    public function validate(array|object $lpa): bool;
-}
