+++ service-api/app/src/App/src/Service/Lpa/SiriusLpaManager.php
@@ -13,7 +13,6 @@ use App\DataAccess\Repository\{InstructionsAndPreferencesImagesInterface,
     ViewerCodesInterface};
 use App\Exception\GoneException;
 use App\Service\Features\FeatureEnabled;
-use App\Service\Lpa\IsValid\IsValidLpa;
 use DateTime;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
@@ -111,7 +110,7 @@ class SiriusLpaManager implements LpaManagerInterface
         }
 
         // Extract and return only LPA's where status is Registered or Cancelled
-        if ($this->isValidLpa->validate((object)$lpaData)) {
+        if (($this->isValidLpa)($lpaData)) {
             return $result;
         }
 
@@ -246,7 +245,7 @@ class SiriusLpaManager implements LpaManagerInterface
             unset($lpaData['original_attorneys']);
 
             //Extract and return only LPA's where status is Registered or Cancelled
-            if ($this->isValidLpa->validate((object)$lpaData)) {
+            if (($this->isValidLpa)($lpaData)) {
                 $result[$item['Id']] = [
                     'user-lpa-actor-token' => $item['Id'],
                     'date'                 => $lpa->getLookupTime()->format('c'),
